% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexPasswordGen.R
\name{password}
\alias{password}
\title{Function to generate secure readable passwords/passphrases}
\usage{
password(no_words = 3, symbols = TRUE, numbers = TRUE, sep = NULL,
  data = Webster)
}
\arguments{
\item{no_words}{- Number of words to be used in the password}

\item{symbols}{- Do we want symbols in the password (TRUE/FALSE)?}

\item{numbers}{- Do we want numbers in the password (TRUE/FALSE)? 
If TRUE random numbers between 100 and 999 are selected.}

\item{sep}{- Specify the separator between the words. 
If NULL a random symbol is selected.}

\item{data}{- words to be used in the password. In default mode the words from
\href{http://www.gutenberg.org/ebooks/29765}{The Project Gutenberg EBook 
of Webster's Unabridged Dictionary, by Various} are used (93077 words).}
}
\description{
Function to generate secure readable passwords/passphrases
}
\examples{
library(PasswordGen)

password(no_words = 7, numbers = FALSE, symbols = FALSE, sep = ' ')
password(no_words = 3, numbers = FALSE, symbols = FALSE, sep = '_')
password(no_words = 3, numbers = FALSE, symbols = TRUE, sep = '_')
password(no_words = 4, numbers = TRUE, symbols = FALSE, sep = '-')
password(1)
password(2)
password()

}

