% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_Package.R
\docType{package}
\name{PasswordGen}
\alias{PasswordGen}
\alias{PasswordGen-package}
\title{PasswordGen: A package for generatin random secure and readable password 
and passphrases using the dictionary method}
\description{
PasswordGen: A package for generatin random secure and readable password 
and passphrases using the dictionary method
}

