% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Webster}
\alias{Webster}
\title{Webster}
\format{An object of class \code{character} of length 93077.}
\source{
The Project Gutenberg EBook of Webster's Unabridged Dictionary, by Various
\url{http://www.gutenberg.org/ebooks/29765}
}
\usage{
Webster
}
\description{
A list extracted from Webster's Unabridged Dictionary
}
\details{
\tabular{rl}{
Title \tab Webster's Unabridged Dictionary\cr
Author \tab Various\cr
Release Date \tab August 22, 2009 [EBook #29765]\cr
Language \tab English
}
}
\keyword{datasets}
