% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexPasswordGen.R
\name{ExtractWords}
\alias{ExtractWords}
\title{Function to extract words from a .txt file}
\usage{
ExtractWords(file, words.min.length = 3)
}
\arguments{
\item{file}{.txt file (say, your favorite book)}

\item{words.min.length}{The minimum legth of the words to be extracted. 
Default value is 3.}
}
\description{
Function to extract words from a .txt file. For example a book from 
Project Gutenberg \url{http://www.gutenberg.org/wiki/Main_Page}.
}
\examples{
\dontrun{
library(PasswordGen)

# Specify the name of a file that you already saved in your workind directory
mybook <- "Project Gutenberg's The Adventures of Sherlock Holmes, 
by Arthur Conan Doyle.txt"

# Extract the unique words
mywords <- ExtractWords(file = mybook, words.min.length = 3)

# Generate password using my words from Sherlock Holmes book.
password(data = mywords)

password(no_words = 7, numbers = FALSE, symbols = FALSE, sep = ' ',
         data = mywords)
} 
}
